requirements/develop.txt
Django==2.1
djangorestframework==3.9.2
pkg-resources==0.0.0
psycopg2==2.7.1
psycopg2-binary==2.7.7
pytz==2018.9
Unipath==1.1
